"learn" modified to "studies"
