+++ src/main/java/pl/roclawski/bartek/spring/studies/repository/ClientRepository.java
@@ -1,4 +1,4 @@
-package pl.roclawski.bartek.spring.learn.repository;
+package pl.roclawski.bartek.spring.studies.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
